add file name configuration from command line or environment for package gcfg
