+++ os/gcfg/gcfg_z_unit_basic_test.go
@@ -12,7 +12,6 @@ import (
 	"os"
 	"testing"
 
-	"github.com/gogf/gf/debug/gdebug"
 	"github.com/gogf/gf/os/gtime"
 
 	"github.com/gogf/gf/encoding/gjson"
@@ -229,71 +228,6 @@ func Test_SetFileName(t *testing.T) {
 	})
 }
 
-func Test_Instance(t *testing.T) {
-	config := `
-array = [1.0, 2.0, 3.0]
-v1 = 1.0
-v2 = "true"
-v3 = "off"
-v4 = "1.234"
-
-[redis]
-  cache = "127.0.0.1:6379,1"
-  disk = "127.0.0.1:6379,0"
-
-`
-	gtest.C(t, func(t *gtest.T) {
-		path := gcfg.DefaultConfigFile
-		err := gfile.PutContents(path, config)
-		t.Assert(err, nil)
-		defer func() {
-			t.Assert(gfile.Remove(path), nil)
-		}()
-
-		c := gcfg.Instance()
-		t.Assert(c.Get("v1"), 1)
-		t.AssertEQ(c.GetInt("v1"), 1)
-		t.AssertEQ(c.GetInt8("v1"), int8(1))
-		t.AssertEQ(c.GetInt16("v1"), int16(1))
-		t.AssertEQ(c.GetInt32("v1"), int32(1))
-		t.AssertEQ(c.GetInt64("v1"), int64(1))
-		t.AssertEQ(c.GetUint("v1"), uint(1))
-		t.AssertEQ(c.GetUint8("v1"), uint8(1))
-		t.AssertEQ(c.GetUint16("v1"), uint16(1))
-		t.AssertEQ(c.GetUint32("v1"), uint32(1))
-		t.AssertEQ(c.GetUint64("v1"), uint64(1))
-
-		t.AssertEQ(c.GetVar("v1").String(), "1")
-		t.AssertEQ(c.GetVar("v1").Bool(), true)
-		t.AssertEQ(c.GetVar("v2").String(), "true")
-		t.AssertEQ(c.GetVar("v2").Bool(), true)
-
-		t.AssertEQ(c.GetString("v1"), "1")
-		t.AssertEQ(c.GetFloat32("v4"), float32(1.234))
-		t.AssertEQ(c.GetFloat64("v4"), float64(1.234))
-		t.AssertEQ(c.GetString("v2"), "true")
-		t.AssertEQ(c.GetBool("v2"), true)
-		t.AssertEQ(c.GetBool("v3"), false)
-
-		t.AssertEQ(c.Contains("v1"), true)
-		t.AssertEQ(c.Contains("v2"), true)
-		t.AssertEQ(c.Contains("v3"), true)
-		t.AssertEQ(c.Contains("v4"), true)
-		t.AssertEQ(c.Contains("v5"), false)
-
-		t.AssertEQ(c.GetInts("array"), []int{1, 2, 3})
-		t.AssertEQ(c.GetStrings("array"), []string{"1", "2", "3"})
-		t.AssertEQ(c.GetArray("array"), []interface{}{1, 2, 3})
-		t.AssertEQ(c.GetInterfaces("array"), []interface{}{1, 2, 3})
-		t.AssertEQ(c.GetMap("redis"), map[string]interface{}{
-			"disk":  "127.0.0.1:6379,0",
-			"cache": "127.0.0.1:6379,1",
-		})
-		t.AssertEQ(c.FilePath(), gfile.Pwd()+gfile.Separator+path)
-
-	})
-}
-
 func TestCfg_New(t *testing.T) {
 	gtest.C(t, func(t *gtest.T) {
 		os.Setenv("GF_GCFG_PATH", "config")
@@ -443,26 +377,6 @@ func TestCfg_Get(t *testing.T) {
 		t.Assert(err, nil)
 		t.Assert(name.Name, "gf")
 		t.Assert(c.GetFloats("floats") == nil, false)
-	})
-}
-
-func TestCfg_Instance(t *testing.T) {
-	gtest.C(t, func(t *gtest.T) {
-		t.Assert(gcfg.Instance("gf") != nil, true)
-	})
-	gtest.C(t, func(t *gtest.T) {
-		pwd := gfile.Pwd()
-		gfile.Chdir(gdebug.TestDataPath())
-		defer gfile.Chdir(pwd)
-		t.Assert(gcfg.Instance("c1") != nil, true)
-		t.Assert(gcfg.Instance("c1").Get("my-config"), "1")
-		t.Assert(gcfg.Instance("folder1/c1").Get("my-config"), "2")
-	})
-	gtest.C(t, func(t *gtest.T) {
-		pwd := gfile.Pwd()
-		gfile.Chdir(gdebug.TestDataPath("folder1"))
-		defer gfile.Chdir(pwd)
-		t.Assert(gcfg.Instance("c2").Get("my-config"), 2)
 	})
 }
