+++ os/gcfg/testdata/envfile/c6.json
@@ -1,0 +1,2 @@
+
+{"my-config": 6}
