+++ os/gcfg/testdata/envpath/c3.toml
@@ -1,0 +1,2 @@
+
+my-config = "3"
