+++ os/gcfg/testdata/envpath/c4.json
@@ -1,0 +1,2 @@
+
+{"my-config": 4}
